requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
gunicorn==23.0.0
python-multipart==0.0.12

--index-url https://www.paddlepaddle.org.cn/packages/stable/cpu/
paddlepaddle==3.2.0
Pillow==11.0.0
opencv-python-headless==4.10.0.84

openai==1.54.0
httpx==0.27.2
pydantic-settings==2.6.1

tenacity==9.0.0
python-dotenv==1.0.1
python-magic==0.4.27
pydantic==2.10.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
